requirements.txt
flask==2.3.3
openai-whisper==20250625
yt-dlp==2023.7.6
groq==0.4.0
python-dotenv==1.0.0
psutil==5.9.5
gunicorn==20.1.0
httpx==0.23.0